docs: refactor examples to use syncmanager instead of persistence adapters
