+++ examples/supabase/src/models/Todos.ts
@@ -1,22 +1,17 @@
-import { Collection, combinePersistenceAdapters, createLocalStorageAdapter } from 'signaldb'
+import { Collection, createLocalStorageAdapter } from 'signaldb'
 import maverickjsReactivityAdapter from 'signaldb-plugin-maverickjs'
-import createSupabasePersistenceAdapter from '../utils/createSupabasePersistenceAdapter'
+import syncManager from '../system/syncManager'
 
 const Todos = new Collection<{ id: string, text: string, completed: boolean }>({
-  memory: [],
   reactivity: maverickjsReactivityAdapter,
-  persistence: combinePersistenceAdapters(
-    createLocalStorageAdapter('todos'),
-    createSupabasePersistenceAdapter<{
-      id: string,
-      text: string,
-      completed: boolean,
-    }, string>('todos'),
-  ),
+  persistence: createLocalStorageAdapter('todos-supabase'),
 })
 Todos.on('persistence.error', (error) => {
   // eslint-disable-next-line no-console
   console.error('persistence.error', error)
 })
+
+syncManager.addCollection(Todos, { name: 'todos' })
+void syncManager.syncAll()
 
 export default Todos
